refactor(web): colocate the turn-id parser with newTurnId, options object for reconcileFetchedMessages (LUM-3470)
turnStartMsFromId parses the format newTurnId owns, so it moves beside
it in send-message-utils; a format change now touches one file, and the
new round-trip test pins the pair together. bucketTurnAge stays in
diagnostics beside bucketMessagesAdded, the tag-banding home.

reconcileFetchedMessages was up to eight positional parameters; the
trailing config (trigger, authoritative, issuedGeneration) becomes an
options object so call sites read as what they are.
