+++ clients/web/src/lib/diagnostics.test.ts
@@ -1,10 +1,6 @@
 import { describe, expect, test } from "bun:test";
 
-import {
-  bucketMessagesAdded,
-  bucketTurnAge,
-  turnStartMsFromId,
-} from "@/lib/diagnostics";
+import { bucketMessagesAdded, bucketTurnAge } from "@/lib/diagnostics";
 
 describe("bucketMessagesAdded", () => {
   // Buckets must stay low-cardinality so the values are aggregable
@@ -46,21 +42,6 @@ describe("bucketMessagesAdded", () => {
     expect(bucketMessagesAdded(Number.NaN)).toBe("0");
     expect(bucketMessagesAdded(Number.POSITIVE_INFINITY)).toBe("0");
     expect(bucketMessagesAdded(Number.NEGATIVE_INFINITY)).toBe("0");
-  });
-});
-
-describe("turnStartMsFromId", () => {
-  test("extracts the send stamp from a client turn id", () => {
-    expect(turnStartMsFromId("turn-1787774471939-au3bcw")).toBe(1787774471939);
-  });
-
-  test("returns null for null and foreign id shapes", () => {
-    expect(turnStartMsFromId(null)).toBe(null);
-    expect(turnStartMsFromId("")).toBe(null);
-    expect(turnStartMsFromId("some-server-id")).toBe(null);
-    expect(turnStartMsFromId("turn-abc-xyz")).toBe(null);
-    // A short digit run is not an epoch-milliseconds stamp.
-    expect(turnStartMsFromId("turn-123-xyz")).toBe(null);
   });
 });
