+++ clients/web/src/lib/diagnostics.ts
@@ -240,24 +240,6 @@ export function bucketMessagesAdded(count: number): string {
   return "6+";
 }
 
-/**
- * The milliseconds embedded in a client turn id (`turn-<ms>-<suffix>`), or
- * null for any other shape. The embedded stamp is the send time, so
- * `now - turnStartMs` bounds how long the turn ran before an observer
- * (e.g. the silent-stall rescue) looked at it.
- */
-export function turnStartMsFromId(turnId: string | null): number | null {
-  if (!turnId) {
-    return null;
-  }
-  const match = /^turn-(\d{10,})-/.exec(turnId);
-  if (!match) {
-    return null;
-  }
-  const ms = Number(match[1]);
-  return Number.isFinite(ms) ? ms : null;
-}
-
 /** Bucket a turn's age at rescue time into a low-cardinality tag band. */
 export function bucketTurnAge(ageMs: number | null): string {
   if (ageMs === null || !Number.isFinite(ageMs) || ageMs < 0) {
